requirements.txt
slack_sdk==3.27.1
python-dotenv==1.0.1
bs4==0.0.2
requests==2.31.0